+++ service/controller/resource/awscnicleaner/create.go
@@ -14,7 +14,7 @@ import (
 	apiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	annotation2 "github.com/giantswarm/aws-operator/v14/pkg/annotation"
+	awsoperatorannotation "github.com/giantswarm/aws-operator/v14/pkg/annotation"
 	"github.com/giantswarm/aws-operator/v14/service/controller/controllercontext"
 	"github.com/giantswarm/aws-operator/v14/service/controller/key"
 )
@@ -146,11 +146,11 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	}
 
 	if key.CiliumPodsCIDRBlock(cluster) != "" {
-		r.logger.Debugf(ctx, "Migrating AWS CNI pod cidr from AWSCluster.Spec.Provider.Pods.CIDRBlock to %q annotation", annotation2.LegacyAwsCniPodCidr)
+		r.logger.Debugf(ctx, "Migrating AWS CNI pod cidr from AWSCluster.Spec.Provider.Pods.CIDRBlock to %q annotation", awsoperatorannotation.LegacyAwsCniPodCidr)
 		if cr.Annotations == nil {
 			cr.Annotations = make(map[string]string)
 		}
-		cr.Annotations[annotation2.LegacyAwsCniPodCidr] = cr.Spec.Provider.Pods.CIDRBlock
+		cr.Annotations[awsoperatorannotation.LegacyAwsCniPodCidr] = cr.Spec.Provider.Pods.CIDRBlock
 
 		r.logger.Debugf(ctx, "Migrating cilium pod cidr from %q annotation to AWSCluster.Spec.Provider.Pods.CIDRBlock", annotation.CiliumPodCidr)
 		// Update pod cidr on AWSCluster CR